+++ binutils/asm/labels.h
@@ -7,7 +7,7 @@ struct label *add_label(char *name,int16_t module,uint16_t addr,int8_t seg);
 void expand_labels();
 void reset_segs_and_module();
 extern uint16_t cur_module;
-extern uint16_t seg_pos[4];
+extern uint32_t seg_pos[4];
 extern uint8_t cur_seg;
 extern unsigned int labels_cur;
 extern unsigned int labels_allocd;
